feat: refactor hash utility functions and update imports; improve error handling in middleware
