+++ middlewares/ExceptionMiddleware.js
@@ -1,7 +1,14 @@
 export const errorHandler = (err, req, res, next) => {
-    const status = 500;
+    console.error(err);
+
+    if (res.headersSent) {
+        return next(err); // Se headers já foram enviados, delega o erro adiante
+    }
+
+    const status = err.status || 500;
     const msg = "Erro interno no servidor";
     const detalhes = err.message;
+
     res.status(status).json({
         status: status,
         msg: msg,
