+++ swagger-output.json
@@ -7,7 +7,7 @@
   },
   "servers": [
     {
-      "url": "http://localhost:5001/"
+      "url": "http://localhost:5000/"
     }
   ],
   "paths": {
