+++ utils/hashUtils.js
